Enable a timeout to be sent to the tunnel
I have separately added support to the tunnel for receiving the timeout in the options.

I have used the same -t switch, but documented it separately in the tunnel command. Perhaps the documentation of the main -t switch should be updated to mention it's also for tunnel timeout?

I needed to extend the timeout of the tunnel for my environment and as it's an arbitrary measurement it seems like a good idea to make it an option.