Implement the --os command line switch
The --os command-line switch didn't appear to be implemented. Obviously this requires support in browseroverfow which I'm submitting separately.

The OS needs to be specified as name:version, e.g. "Windows:XP" or "OS X:Mountain Lion". That's a bit clumsy looking, but it fits with the browser version style and it sort of matches the output from the "browsers" command.

Perhaps it could be supported with the "Name (Version)" style. If you like the sound of that I would be happy to implement that for both OS and browser.